backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
pymongo==4.6.1

# AI and ML
google-generativeai==0.3.2

# File storage and media
cloudinary==1.36.0
Pillow==10.1.0

# Animation and mathematical libraries
manim==0.18.0
numpy==1.24.3
matplotlib==3.7.2

# Security and authentication
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities and environment
python-dotenv==1.0.0
Werkzeug==3.0.1

# Validation and serialization
marshmallow==3.20.2
email-validator==2.1.0

# HTTP requests
requests==2.31.0

# Async support
asyncio==3.4.3

# Development and testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production dependencies
gunicorn==21.2.0
supervisor==4.2.5

# Monitoring and logging
structlog==23.2.0
